feat: create api server ts, server inly using node:https
