+++ app/sitemap.xml/route.ts
@@ -1,4 +1,4 @@
-import https from "node:https";
+import { getAllPostsForSitemap } from "../../lib/api-server";
 import {
   adaptPostsForSitemap,
   assertFullSitemap,
@@ -14,119 +14,17 @@ import {
 
 // ISR: Vercel caches this response in its CDN edge cache for 1 hour.
 //
-// After first generation: every request served from CDN (<10ms, no Lambda invoked).
+// After first generation: every request is served from CDN (<10ms, no Lambda invoked).
 // After TTL expires: stale version served immediately, regeneration happens in background.
-// If WordPress is down during regen: Vercel keeps serving previous good version automatically.
-// Cold-start / first request: mitigated by build-time pre-generation (see scripts/prewarm-sitemap.mjs)
-// and post-deploy warming triggered by the Vercel deployment hook in GitHub Actions.
+// If WordPress is down during regen: Vercel keeps serving the previous good version automatically.
+// Cold-start / first request: mitigated by post-deploy warming via .github/workflows/prewarm-sitemap.yml.
 export const revalidate = 3600;
-
-// ---------------------------------------------------------------------------
-// fetchGraphQL — bypasses Next.js App Router's RSC fetch instrumentation.
-//
-// Problem: Next.js wraps global fetch() in RSC/Route Handler context with its
-// own caching layer. This instrumented fetch sends headers that Cloudflare
-// (in front of wp.keploy.io) interprets as invalid, returning 502 HTML instead
-// of JSON — even though the same URL works fine from plain Node.js.
-//
-// Fix: use node:https directly. This is a raw TCP connection with no Next.js
-// middleware in the path, identical to what curl or plain `node -e "fetch()"` sends.
-// ---------------------------------------------------------------------------
-function fetchGraphQL(query: string, variables: Record<string, unknown> = {}): Promise<any> {
-  const apiUrl = process.env.WORDPRESS_API_URL || process.env.NEXT_PUBLIC_WORDPRESS_API_URL;
-  if (!apiUrl) throw new Error("WORDPRESS_API_URL is not configured");
-
-  const url = new URL(apiUrl);
-  const body = JSON.stringify({ query, variables });
-
-  return new Promise((resolve, reject) => {
-    const req = https.request(
-      {
-        hostname: url.hostname,
-        port: url.port || 443,
-        path: url.pathname + url.search,
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "Content-Length": Buffer.byteLength(body),
-          "User-Agent": "keploy-blog-sitemap/1.0",
-        },
-      },
-      (res) => {
-        let data = "";
-        res.on("data", (chunk) => (data += chunk));
-        res.on("end", () => {
-          try {
-            const json = JSON.parse(data);
-            if (json.errors) reject(new Error(JSON.stringify(json.errors)));
-            else resolve(json.data);
-          } catch (e) {
-            reject(new Error(`WordPress returned non-JSON (status ${res.statusCode}): ${data.slice(0, 120)}`));
-          }
-        });
-      }
-    );
-    req.on("error", reject);
-    req.write(body);
-    req.end();
-  });
-}
-
-// Paginates through all posts using cursor-based pagination.
-// Identical query to lib/api.ts getAllPosts() but using fetchGraphQL instead of fetch().
-async function getAllPostsForSitemap() {
-  let allEdges: any[] = [];
-  let hasNextPage = true;
-  let endCursor: string | null = null;
-
-  while (hasNextPage) {
-    const data = await fetchGraphQL(
-      `
-      query AllPosts($after: String) {
-        posts(first: 50, after: $after, where: { orderby: { field: DATE, order: DESC } }) {
-          edges {
-            node {
-              slug
-              modified
-              ppmaAuthorName
-              categories {
-                edges {
-                  node {
-                    name
-                    slug
-                  }
-                }
-              }
-              tags {
-                edges {
-                  node {
-                    name
-                  }
-                }
-              }
-            }
-          }
-          pageInfo {
-            hasNextPage
-            endCursor
-          }
-        }
-      }
-      `,
-      { after: endCursor }
-    );
-
-    const edges = data?.posts?.edges ?? [];
-    allEdges = [...allEdges, ...edges];
-    hasNextPage = data?.posts?.pageInfo?.hasNextPage ?? false;
-    endCursor = data?.posts?.pageInfo?.endCursor ?? null;
-  }
-
-  return { edges: allEdges };
-}
 
 export async function GET(): Promise<Response> {
   try {
+    // getAllPostsForSitemap() uses node:https directly (see lib/api-server.ts) to
+    // bypass Next.js App Router's RSC fetch instrumentation, which causes Cloudflare
+    // to return 502 HTML when global fetch() is used in this context.
     const allPostsResult = await getAllPostsForSitemap();
     const posts = adaptPostsForSitemap(allPostsResult);
 
@@ -134,7 +32,8 @@ export async function GET(): Promise<Response> {
     // so Vercel keeps serving the previous good cached version automatically.
     assertFullSitemap(posts);
 
-    // Static routes get lastmod = newest post modification time.
+    // Static routes get lastmod = newest post modification time,
+    // so listing pages reflect when the freshest underlying content changed.
     const latestModified = getLatestModified(posts) ?? new Date().toISOString();
     const staticEntries = STATIC_ROUTES.map((r) => ({
       ...r,
@@ -167,7 +66,8 @@ export async function GET(): Promise<Response> {
     );
 
     // ISR does NOT cache non-2xx responses.
-    // no-store prevents any downstream proxy from caching this degraded response.
+    // no-store prevents any downstream proxy from caching this degraded response,
+    // so crawlers will retry on the next request once WordPress is back.
     return new Response(getStaticFallbackXml(), {
       status: 503,
       headers: {
